+++ bundles/de.uka.ipd.sdq.workflow.mdsd.core/META-INF/MANIFEST.MF
@@ -3,13 +3,16 @@ Bundle-ManifestVersion: 2
 Bundle-Name: Core MDSD Jobs for Palladio Workflow Engine
 Bundle-SymbolicName: de.uka.ipd.sdq.workflow.mdsd.core
 Bundle-Version: 5.2.1.qualifier
-Export-Package: de.uka.ipd.sdq.workflow.mdsd.core
+Export-Package: de.uka.ipd.sdq.workflow.mdsd.core,
+ de.uka.ipd.sdq.workflow.mdsd.core.emf.qvto
 Require-Bundle: org.eclipse.core.runtime;bundle-version="3.26.100",
  org.eclipse.debug.core;bundle-version="3.20.0",
  org.apache.log4j;bundle-version="1.2.19",
  de.uka.ipd.sdq.workflow.launchconfig.core;bundle-version="5.2.1",
  de.uka.ipd.sdq.workflow.logging.console;bundle-version="5.2.1",
- de.uka.ipd.sdq.workflow.mdsd.blackboard;bundle-version="5.2.1"
+ de.uka.ipd.sdq.workflow.mdsd.blackboard;bundle-version="5.2.1",
+ com.google.guava;bundle-version="27.1.0",
+ org.eclipse.m2m.qvt.oml;bundle-version="3.10.7"
 Bundle-Vendor: sdq.ipd.uka.de
 Bundle-RequiredExecutionEnvironment: JavaSE-17
 Bundle-ActivationPolicy: lazy
