+++ tests/de.uka.ipd.sdq.workflow.mdsd.tests/src/de/uka/ipd/sdq/workflow/mdsd/tests/QVTOTests.java
@@ -33,8 +33,8 @@ import de.uka.ipd.sdq.workflow.jobs.UserCanceledException;
 import de.uka.ipd.sdq.workflow.mdsd.blackboard.MDSDBlackboard;
 import de.uka.ipd.sdq.workflow.mdsd.blackboard.ModelLocation;
 import de.uka.ipd.sdq.workflow.mdsd.blackboard.ResourceSetPartition;
-import de.uka.ipd.sdq.workflow.mdsd.emf.qvto.QVTOTransformationJob;
-import de.uka.ipd.sdq.workflow.mdsd.emf.qvto.QVTOTransformationJobConfiguration;
+import de.uka.ipd.sdq.workflow.mdsd.core.emf.qvto.QVTOTransformationJob;
+import de.uka.ipd.sdq.workflow.mdsd.core.emf.qvto.QVTOTransformationJobConfiguration;
 import tools.mdsd.library.standalone.initialization.StandaloneInitializationException;
 import tools.mdsd.library.standalone.initialization.StandaloneInitializerBuilder;
